requirements.txt
python-dotenv~=1.0.1
py-cord~=2.6.1
aiosqlite
audioop-lts
matplotlib~=3.10.1

numpy~=2.2.3